Making job_sort_formula a sched atttribute
