--- src/include/qmgr_node_public.h
@@ -1,68 +1,0 @@
-/*
- * Copyright (C) 1994-2019 Altair Engineering, Inc.
- * For more information, contact Altair at www.altair.com.
- *
- * This file is part of the PBS Professional ("PBS Pro") software.
- *
- * Open Source License Information:
- *
- * PBS Pro is free software. You can redistribute it and/or modify it under the
- * terms of the GNU Affero General Public License as published by the Free
- * Software Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * PBS Pro is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.
- * See the GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Commercial License Information:
- *
- * For a copy of the commercial license terms and conditions,
- * go to: (http://www.pbspro.com/UserArea/agreement.html)
- * or contact the Altair Legal Department.
- *
- * Altair’s dual-license business model allows companies, individuals, and
- * organizations to create proprietary derivative works of PBS Pro and
- * distribute them - whether embedded or bundled with other software -
- * under a commercial license agreement.
- *
- * Use of Altair’s trademarks, including but not limited to "PBS™",
- * "PBS Professional®", and "PBS Pro™" and Altair’s logos is subject to Altair's
- * trademark licensing policies.
- *
- */
-
-/*
- *	This is a list of public node attributes
- *
- *	FORMAT:
- *		attr1,
- * 		attr2,	<--- important the last has a comma after it
- *
- * 	This file will be used for the initialization of an array.  Also,
- *	attributes listed here get printed on a qmgr print operation to be
- *	fed back into qmgr to restore a configuration.
- *
- *
- */
-
-ATTR_NODE_Mom,
-ATTR_NODE_Port,
-ATTR_NODE_state,
-ATTR_NODE_Sharing,
-ATTR_p,
-ATTR_rescavail,
-ATTR_queue,
-ATTR_maxrun,
-ATTR_maxuserrun,
-ATTR_maxgrprun,
-ATTR_comment,
-ATTR_NODE_NoMultiNode,
-ATTR_NODE_resv_enable,
-ATTR_PNames,
-ATTR_NODE_ProvisionEnable,
-ATTR_NODE_current_aoe,
