+++ src/lib/Libattr/master_sched_attr_def.xml
@@ -476,6 +476,22 @@
         <ECL>NULL_VERIFY_VALUE_FUNC</ECL>
         </member_verify_function>
    </attributes>
+   <attributes>   /* job_sort_formula */    
+    <member_name><both>ATTR_job_sort_formula</both></member_name>     <!-- "job_sort_formula" -->
+    <member_at_decode>decode_formula</member_at_decode>
+    <member_at_encode>encode_str</member_at_encode>
+    <member_at_set>set_str</member_at_set>
+    <member_at_comp>comp_str</member_at_comp>
+    <member_at_free>free_str</member_at_free>
+    <member_at_action>validate_job_formula</member_at_action>
+    <member_at_flags><both>MGR_ONLY_SET</both></member_at_flags>
+    <member_at_type><both>ATR_TYPE_STR</both></member_at_type>
+    <member_at_parent>PARENT_TYPE_SCHED</member_at_parent>
+    <member_verify_function>
+    <ECL>NULL_VERIFY_DATATYPE_FUNC</ECL>
+    <ECL>NULL_VERIFY_VALUE_FUNC</ECL>
+    </member_verify_function>
+   </attributes>
 
     <tail>
      <SVR>
