+++ src/lib/Libattr/master_svr_attr_def.xml
@@ -1290,7 +1290,7 @@
 	<member_at_comp>comp_str</member_at_comp>
 	<member_at_free>free_str</member_at_free>
 	<member_at_action>validate_job_formula</member_at_action>
-	<member_at_flags><both>READ_WRITE</both></member_at_flags>
+	<member_at_flags><both>MGR_ONLY_SET</both></member_at_flags>
 	<member_at_type><both>ATR_TYPE_STR</both></member_at_type>
 	<member_at_parent>PARENT_TYPE_SERVER</member_at_parent>
 	<member_verify_function>
