+++ src/lib/Liblog/pbs_messages.c
@@ -411,6 +411,7 @@ char *msg_invalid_partion_in_queue = "Invalid partition in queue";
 char *msg_sched_op_not_permitted = "Operation is not permitted on default scheduler";
 char *msg_sched_part_already_used = "Partition is already associated with other scheduler";
 char *msg_invalid_max_job_sequence_id = "Cannot set max_job_sequence_id < 9999999, or > 999999999999";
+char *msg_jsf_incompatible = "Server's job_sort_formula value is incompatible with sched's";
 
 char *msg_resv_not_empty = "Reservation not empty";
 char *msg_stdg_resv_occr_conflict = "Requested time(s) will interfere with a later occurrence";
@@ -596,6 +597,7 @@ struct pbs_err_to_txt pbs_err_to_txt[] = {
 	{PBSE_SCHED_OP_NOT_PERMITTED, &msg_sched_op_not_permitted},
 	{PBSE_SCHED_PARTITION_ALREADY_EXISTS, &msg_sched_part_already_used},
 	{PBSE_INVALID_MAX_JOB_SEQUENCE_ID, &msg_invalid_max_job_sequence_id},
+	{PBSE_SVR_SCHED_JSF_INCOMPAT, &msg_jsf_incompatible},
 	{ 0, NULL }		/* MUST be the last entry */
 };
